Extracted Transition.inlineOperations function
